feat(component): add ThemeSwitcher button
